+++ bin/runtime-common/src/lib.rs
@@ -24,10 +24,10 @@ use xcm::v3::NetworkId;
 
 pub mod messages;
 pub mod messages_api;
-pub mod messages_batch_extension;
 pub mod messages_benchmarking;
 pub mod messages_extension;
 pub mod parachains_benchmarking;
+pub mod refund_relayer_extension;
 
 mod messages_generation;
